requirement.txt
streamlit
gensim
python_glove
pickle